move updating wallet balance to seperate function
