+++ src/components/partials/Sidebar.js
@@ -1,399 +1,12 @@
 import React, { Component } from "react";
-import { roundAmount } from "../../utils/utils.js";
 import ReactTooltip from "react-tooltip";
 import { CopyToClipboard } from "react-copy-to-clipboard";
-import ContactName from "./ContactName";
+import Transactions from "./Transactions";
+import Contacts from "./Contacts";
 import * as crypto from "crypto";
 
 const safex = window.require("safex-nodejs-libwallet");
-const { shell } = window.require("electron");
 const fileDownload = require("react-file-download");
-
-class Transactions extends Component {
-  constructor(props) {
-    super(props);
-    this.state = { tx_page: 0 };
-    this.totalTxPages = 0;
-  }
-
-  firstTxPage = () => {
-    this.setState({ tx_page: 0 });
-  };
-
-  previousTxPage = () => {
-    if (this.state.tx_page >= 1) {
-      this.setState(prevState => ({ tx_page: prevState.tx_page - 1 }));
-    }
-  };
-
-  nextTxPage = () => {
-    if (this.state.tx_page < this.totalTxPages - 1) {
-      this.setState(prevState => ({ tx_page: prevState.tx_page + 1 }));
-    }
-  };
-
-  lastTxPage = () => {
-    this.setState({ tx_page: this.totalTxPages - 1 });
-  };
-
-  externalLink = txid => {
-    if (process.env.NODE_ENV === "development") {
-      shell.openExternal("http://178.128.89.114/search?value=" + txid);
-    } else {
-      shell.openExternal("http://explore.safex.io/search?value=" + txid);
-    }
-  };
-
-  render() {
-    const { tx_page } = this.state;
-    const { itemsPerPage, history } = this.props;
-    this.totalTxPages = Math.ceil(history.length / itemsPerPage);
-    let options = {
-      weekday: "long",
-      year: "numeric",
-      month: "long",
-      day: "numeric",
-      hour: "numeric",
-      minute: "numeric",
-      hour12: true
-    };
-
-    return (
-      <div>
-        {this.props.history.length ? (
-          history
-            .slice(
-              tx_page * itemsPerPage,
-              tx_page * itemsPerPage + itemsPerPage
-            )
-            .map((txInfo, i) => {
-              return (
-                <div className="history-item" key={i}>
-                  <div className="row">
-                    <div className="col-xs-5 item-section">
-                      <p className={txInfo.pending ? "hidden" : ""}>
-                        <img
-                          src={
-                            txInfo.direction === "in"
-                              ? "images/arrow-down.png"
-                              : "images/arrow-up.png"
-                          }
-                          className="arrow-img"
-                          alt="arrow-img"
-                        />
-                        <span
-                          className={
-                            txInfo.direction === "in" ? "green-text" : ""
-                          }
-                        >
-                          {txInfo.direction === "in" ? "Received" : "Sent"}
-                        </span>
-                      </p>
-                      <p className={txInfo.pending ? "yellow-text" : "hidden"}>
-                        {txInfo.pending}
-                      </p>
-                      {roundAmount(txInfo.tokenAmount) === 0 ? (
-                        <span
-                          className={
-                            txInfo.direction === "in"
-                              ? "green-text amount"
-                              : "amount"
-                          }
-                        >
-                          SFX {roundAmount(txInfo.amount)}
-                        </span>
-                      ) : (
-                        <span
-                          className={
-                            txInfo.direction === "in"
-                              ? "green-text amount"
-                              : "amount"
-                          }
-                        >
-                          SFT {roundAmount(txInfo.tokenAmount)}
-                        </span>
-                      )}
-                    </div>
-                    <div className="col-xs-7 item-section">
-                      <p className="text-right">
-                        {"" +
-                          new Date(txInfo.timestamp * 1000).toLocaleDateString(
-                            "en-US",
-                            options
-                          )}
-                      </p>
-                      <p className={txInfo.direction === "in" ? "hidden" : ""}>
-                        Fee: SFX {roundAmount(txInfo.fee)}
-                      </p>
-                    </div>
-                  </div>
-
-                  <div className="tx-id-wrap">
-                    <span>Transaction ID:</span>
-                    <p
-                      data-tip
-                      data-for="link-tooptip"
-                      className="tx-id"
-                      onClick={this.externalLink.bind(this, txInfo.id)}
-                    >
-                      {txInfo.id}
-                    </p>
-                    <ReactTooltip id="link-tooptip">
-                      <p>
-                        Show this transaction on{" "}
-                        <span className="blue-text">
-                          Safex Blockchain Explorer
-                        </span>
-                      </p>
-                    </ReactTooltip>
-                  </div>
-                </div>
-              );
-            })
-        ) : (
-          <h5>No Transaction History</h5>
-        )}
-
-        {this.props.history.length ? (
-          <div className="pagination">
-            <button
-              data-tip
-              data-for="first-tooptip"
-              className="first-page button-shine"
-              onClick={this.firstTxPage}
-            >
-              <span>{"<<"}</span>
-            </button>
-            <ReactTooltip id="first-tooptip">
-              <p>First Page</p>
-            </ReactTooltip>
-            <button className="button-shine" onClick={this.previousTxPage}>
-              previous
-            </button>
-            <strong>
-              page: {tx_page + 1} / {this.totalTxPages}
-            </strong>{" "}
-            <button className="button-shine" onClick={this.nextTxPage}>
-              next
-            </button>
-            <button
-              data-tip
-              data-for="last-tooptip"
-              className="last-page button-shine"
-              onClick={this.lastTxPage}
-            >
-              <span>{">>"}</span>
-            </button>
-            <ReactTooltip id="last-tooptip">
-              <p>Last Page</p>
-            </ReactTooltip>
-          </div>
-        ) : (
-          <div className="hidden" />
-        )}
-      </div>
-    );
-  }
-}
-
-class Contacts extends Component {
-  constructor(props) {
-    super(props);
-    this.state = { contact_page: 0, editing: false };
-    this.totalContactPages = 0;
-  }
-
-  firstContactPage = () => {
-    this.setState({ contact_page: 0 });
-  };
-
-  previousContactPage = () => {
-    if (this.state.contact_page >= 1) {
-      this.setState(prevState => ({
-        contact_page: prevState.contact_page - 1
-      }));
-    }
-  };
-
-  nextContactPage = () => {
-    if (this.state.contact_page < this.totalContactPages - 1) {
-      this.setState(prevState => ({
-        contact_page: prevState.contact_page + 1
-      }));
-    }
-  };
-
-  lastContactPage = () => {
-    this.setState({ contact_page: this.totalContactPages - 1 });
-  };
-
-  removeContact = rowID => {
-    this.props.setOpenDeleteModal();
-    localStorage.setItem("rowID", rowID);
-  };
-
-  editContact = (address, paymentID, name, rowID) => {
-    let wallet = this.props.walletMeta;
-    wallet.addressBook_DeleteRow(rowID);
-    setTimeout(() => {
-      wallet.addressBook_AddRow(address, paymentID, name);
-      wallet.store();
-      console.log("Wallet stored");
-      this.props.setWalletHistory();
-      localStorage.removeItem("rowID");
-    }, 100);
-    console.log("Wallet edited");
-  };
-
-  render() {
-    const { contact_page } = this.state;
-    const { itemsPerContactPage, contacts } = this.props;
-    this.totalContactPages = Math.ceil(contacts.length / itemsPerContactPage);
-
-    return (
-      <div>
-        {this.props.contacts.length ? (
-          contacts
-            .slice(
-              contact_page * itemsPerContactPage,
-              contact_page * itemsPerContactPage + itemsPerContactPage
-            )
-            .map((contact, i) => {
-              return (
-                <div className="contact-item" key={i}>
-                  <div className="contact-left">
-                    {/* <ContactName
-                      address={contact.address}
-                      paymentId={contact.paymentId}
-                      contactName={contact.description}
-                      editContact={this.editContact}
-                      rowID={contact.rowID}
-                    /> */}
-                    <p className="name">{contact.description}</p>
-                    <p className="payment-id">Payment ID:</p>
-                  </div>
-                  <ReactTooltip id="edit-tooptip">
-                    <p>Edit Contact</p>
-                  </ReactTooltip>
-
-                  <div className="contact-center">
-                    <p className="address">{contact.address}</p>
-                    <p className="payment-id">{contact.paymentId}</p>
-                  </div>
-
-                  <div className="contact-right">
-                    <CopyToClipboard
-                      text={contact.address}
-                      onCopy={this.props.onCopy.bind(
-                        this,
-                        "Copied to clipboard",
-                        3000
-                      )}
-                      data-tip
-                      data-for="copy-tooptip"
-                      className="button-shine copy-btn"
-                    >
-                      <button>
-                        <img src="images/copy.png" alt="copy" />
-                      </button>
-                    </CopyToClipboard>
-                    <ReactTooltip id="copy-tooptip">
-                      <p>Copy Address</p>
-                    </ReactTooltip>
-                    <button
-                      onClick={this.removeContact.bind(this, contact.rowID)}
-                      data-tip
-                      data-for="remove-tooptip"
-                      className="button-shine"
-                    >
-                      <img src="images/trash.png" alt="trash" />
-                    </button>
-                    <ReactTooltip id="remove-tooptip">
-                      <p>Remove Contact</p>
-                    </ReactTooltip>
-                    <button
-                      onClick={this.props.setOpenSendModal.bind(
-                        this,
-                        0,
-                        contact.address,
-                        contact.paymentId,
-                        contact.description
-                      )}
-                      data-tip
-                      data-for="send-cash-tooltip"
-                      className="button-shine"
-                    >
-                      <img src="images/send-cash.png" alt="send-cash" />
-                    </button>
-                    <ReactTooltip id="send-cash-tooltip">
-                      <p>Send Safex Cash To This Address</p>
-                    </ReactTooltip>
-                    <button
-                      onClick={this.props.setOpenSendModal.bind(
-                        this,
-                        1,
-                        contact.address,
-                        contact.paymentId,
-                        contact.description
-                      )}
-                      data-tip
-                      data-for="send-token-tooltip"
-                      className="button-shine"
-                    >
-                      <img src="images/send-token.png" alt="send-token" />
-                    </button>
-                    <ReactTooltip id="send-token-tooltip">
-                      <p>Send Safex Token To This Address</p>
-                    </ReactTooltip>
-                  </div>
-                </div>
-              );
-            })
-        ) : (
-          <h5>No Contacts</h5>
-        )}
-
-        {this.props.contacts.length ? (
-          <div className="pagination">
-            <button
-              data-tip
-              data-for="first-tooptip"
-              className="first-page button-shine"
-              onClick={this.firstContactPage}
-            >
-              <span>{"<<"}</span>
-            </button>
-            <ReactTooltip id="first-tooptip">
-              <p>First Page</p>
-            </ReactTooltip>
-            <button className="button-shine" onClick={this.previousContactPage}>
-              previous
-            </button>
-            <strong>
-              page: {contact_page + 1} / {this.totalContactPages}
-            </strong>{" "}
-            <button className="button-shine" onClick={this.nextContactPage}>
-              next
-            </button>
-            <button
-              data-tip
-              data-for="last-tooptip"
-              className="last-page button-shine"
-              onClick={this.lastContactPage}
-            >
-              <span>{">>"}</span>
-            </button>
-            <ReactTooltip id="last-tooptip">
-              <p>Last Page</p>
-            </ReactTooltip>
-          </div>
-        ) : (
-          <div className="hidden" />
-        )}
-      </div>
-    );
-  }
-}
 
 export default class Sidebar extends Component {
   constructor(props) {
@@ -753,7 +366,6 @@ export default class Sidebar extends Component {
                 itemsPerContactPage={3}
                 setOpenAlert={this.props.setOpenAlert}
                 onCopy={this.props.onCopy}
-                setWalletData={this.props.setWalletData}
                 setWalletHistory={this.props.setWalletHistory}
                 setOpenSendModal={this.props.setOpenSendModal}
                 setOpenDeleteModal={this.props.setOpenDeleteModal}
